fix(lib): Allow action nodes to end with known-safe functions
Fix a bug in which even safe sanitizers had their actions re-terminated
with the default sanitizer.

Fix another bug in which the `template.FuncMap` generated for the
sanitizers failed to dereference `Format` on the sanitizer and instead
attempted to pass the sanitizer value itself as a function, which
panicked at runtime.

Fix another bug in which the sanitizers never got added to the root
template `FuncMap`.